feat(车规): 静态分析 0-warning + ASIL/MISRA 安全文档 (P5/P6/P8)
P5 编译开关 — -Wall -Wextra -Werror -Wshadow -Wconversion -Wsign-conversion
  -Wold-style-cast -Wcast-align -Wnon-virtual-dtor 在 can-processor/CMakeLists.txt
  落实，修完所有 warning

P6 cppcheck + clang-tidy 静态分析配置
  - .cppcheck: warning/style/performance/portability + inline-suppr
  - .clang-tidy: 聚焦 cert-* bugprone-* performance-* clang-analyzer-*
    噪声规则 (modernize/readability/cppcoreguidelines) 关闭
    关键安全检查 cert/bugprone-use-after-move/analyzer-core.NullDereference
    等 17 条设为 WarningsAsErrors
  - 修复 cppcheck 误报: NVI 析构改调非虚 closeImpl_()，suppress 注释在 .h
    dtor 声明行前
  - 修复 clang-tidy 真实缺陷:
    - shm_display.cpp lround 替换 (int)(hz*10+0.5) 浮点四舍五入 bug
    - vehicle_logic.cpp / can_signal_monitor.cpp 显式 static_cast<float>
      避免 int→float narrowing
    - AlarmRuntime 析构 delete[] m_states (subagent FMEA P0 整改项)
    - socketcan/sim_socket close() NVI 转发补回 (被前序误删)
    - easily-swappable-parameters 全部 NOLINTBEGIN/END 抑制 (CAN 帧格式
      接口固定)
    - event_bus subscribe handler NOLINT (std::function move 代价等同 copy)
  - 验证: cppcheck 0 / clang-tidy 0 / build OK

P8 ASIL/MISRA 合规 + 安全分析文档
  - docs/MISRA-COMPLIANCE.md (539 行)
    17 条 MISRA C++:2008 + CERT + C++CG 实施表 (含行号引用)
    工具链配置 + CI 集成 + 7 条未实现项路线图
    审计 Checklist 10 项
  - docs/SAFETY-ANALYSIS.md (486 行)
    12 个信号 ASIL 分解 (车速/报警/安全带 = ASIL-B)
    18 条 FMEA + RPN 排序
    9 条安全机制现状 + 6 条待实施
    ASCII 架构图 + 残余风险评审
    PMHF 估算 < 10^-7/h (满足 ASIL-B)

+++ src/layer2/vehicle_logic.cpp
@@ -64,7 +64,7 @@ void VehicleLogic::onSocUpdate(float soc) {
             count++;
         }
     }
-    m_socSmoothed = (count > 0) ? (sum / count) : soc;
+    m_socSmoothed = (count > 0) ? (sum / static_cast<float>(count)) : soc;
     m_soc = soc;
 
     Event e2{/*key=*/"bat_soc", /*value=*/m_socSmoothed, /*prev_value=*/m_lastSoc,
